Fix UI session runtime bundle split
